// app/types/order.ts
export enum Margin_Type {
  Cross = "Cross",
  Isolated = "Isolated",
}

export enum Order_Type {
  Limit = "limit",
  Market = "Market",
}

export interface OrderToPlace {
  type: Order_Type;
  price: string;
  stopLossPrice: string;
  takeProfitPrice: string;
  unit: string;
  amount: string;
  isBuyOrder: boolean;
  hasTime: boolean;
  expiration: any;
  leverage: number;
  margin: Margin_Type;
}